# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a VideoWeb API",
        "title": "VideoWeb",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/fans/follows": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "关注其他用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要关注的用户ID",
                        "name": "FID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/fans/unfollows": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "取消关注其他用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要取消关注的用户ID",
                        "name": "FID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/favorites/create": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "创建收藏夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "收藏夹名称",
                        "name": "FName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "公开",
                            "私密"
                        ],
                        "type": "string",
                        "description": "是否私密",
                        "name": "IsPrivate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "Description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/favorites/delete": {
            "delete": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "删除收藏夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "收藏夹名称",
                        "name": "FName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/favorites/modify": {
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "修改收藏夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "收藏夹ID",
                        "name": "FavoriteID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要修改的收藏夹名称",
                        "name": "newName",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "公开",
                            "私密"
                        ],
                        "type": "string",
                        "description": "是否私密",
                        "name": "IsPrivate",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "Description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":\"200\",\"data\":\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}